# app/Models/User.php
<?php

namespace App\Models;

// use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;
use Laravel\Sanctum\HasApiTokens;
use Stripe\Terminal\Location;
use Tymon\JWTAuth\Contracts\JWTSubject;

class User extends Authenticatable implements JWTSubject
{
    use HasApiTokens, HasFactory, Notifiable;


    public function favorites()
{
    return $this->belongsToMany(Product::class, 'favorites')->withTimestamps();
}


    public function roles(){
        return $this->belongsToMany(Role::class);
    }

// Check if the user has a specific role

    public function hasRole($role)
{
    return $this->roles()->where('name', $role)->exists();
}





 // Check if the user has any of the specified roles
 public function hasAnyRole(array $roles)
 {
     foreach ($roles as $role) {
         if ($this->hasRole($role)) {
             return true;
         }
     }
     return false;
 }

public function permissions()
{
    return $this->roles()->permissions(); // Access permissions via roles
}

public function products(){
    return $this->hasMany(Product::class);
}

public function orders(){
    return $this->hasMany(Order::class);
}



    /**
     * The attributes that are mass assignable.
     *
     * @var array<int, string>
     */
    protected $fillable = [
        'name',
        'email',
        'password',
    ];

    /**
     * The attributes that should be hidden for serialization.
     *
     * @var array<int, string>
     */
    protected $hidden = [
        'password',
        'remember_token',
        'created_at',
        'updated_at'
    ];

    /**
     * The attributes that should be cast.
     *
     * @var array<string, string>
     */
    protected $casts = [
        'email_verified_at' => 'datetime',
        'password' => 'hashed',
    ];

    public function getJWTIdentifier()
    {
        return $this->getKey();
    }

    public function getJWTCustomClaims()
    {
        return [];
    }


   
}